+++ src/AbstractApi.php
@@ -53,7 +53,7 @@ abstract class AbstractApi
     /**
      * PHP API version
      */
-    const PHP_API_VERSION = '3.1.2';
+    const PHP_API_VERSION = '4.0.0';
 
     /**
      * Event dispatcher