+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added User-Agent in API call headers
+- Added new properties for ReleaseReservationResponse class
 
 ### Changed